Add a way to escape enemies
When we know we are going to loose a fight. Because right now the ants often end up wandering alone, there is no route planning.
This doesn't allow to live forever but it allows to live for a few more turns and hopefully we can manage to grab a bit more food before dying.